idk/requirements.txt
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask==2.0.3
Werkzeug==2.0.3
opencv_python==4.7.0.68
numpy==1.24.2
gunicorn==20.1.0
eventlet==0.30.2
psycopg2-binary
pygame
flask_login
gtts
scikit-learn==1.1
mediapipe==0.10.9
pillow